Move out the common shared code
