+++ src/block/wasm32.rs
@@ -11,12 +11,6 @@ use core::{
 pub struct Block(v128);
 
 impl Block {
-    const _ASSERTION: () = {
-        if core::mem::size_of::<Self>() % core::mem::size_of::<usize>() != 0 {
-            panic!("vector is not a multiple size of usize");
-        }
-    };
-
     pub const USIZE_COUNT: usize = core::mem::size_of::<Self>() / core::mem::size_of::<usize>();
     pub const NONE: Self = Self::from_usize_array([0; Self::USIZE_COUNT]);
     pub const ALL: Self = Self::from_usize_array([core::usize::MAX; Self::USIZE_COUNT]);
@@ -100,51 +94,5 @@ impl PartialEq for Block {
     #[inline]
     fn eq(&self, other: &Self) -> bool {
         !v128_any_true(v128_xor(self.0, other.0))
-    }
-}
-
-impl Eq for Block {}
-
-impl PartialOrd for Block {
-    #[inline]
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        let a = self.into_usize_array();
-        let b = other.into_usize_array();
-        for i in 0..Self::USIZE_COUNT {
-            match a[i].cmp(&b[i]) {
-                Ordering::Equal => continue,
-                cmp => return Some(cmp),
-            }
-        }
-        Some(Ordering::Equal)
-    }
-}
-
-impl Ord for Block {
-    #[inline]
-    fn cmp(&self, other: &Self) -> Ordering {
-        let a = self.into_usize_array();
-        let b = other.into_usize_array();
-        for i in 0..Self::USIZE_COUNT {
-            match a[i].cmp(&b[i]) {
-                Ordering::Equal => continue,
-                cmp => return cmp,
-            }
-        }
-        Ordering::Equal
-    }
-}
-
-impl Default for Block {
-    #[inline]
-    fn default() -> Self {
-        Self::NONE
-    }
-}
-
-impl Hash for Block {
-    #[inline]
-    fn hash<H: Hasher>(&self, hasher: &mut H) {
-        self.into_usize_array().hash(hasher)
     }
 }
